pyproject.toml
[tool.poetry]
name = "my-gateway"
version = "0.1.0"
description = ""
authors = ["MauroMartinsJr <maurojosedmjr@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.3"
ruff = "^0.4.2"
uvicorn = "^0.29.0"
requests = "^2.31.0"
pytz = "^2024.1"
pyjwt = "^2.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
